+++ src/ExtensionActivator.ts
@@ -8,7 +8,7 @@ import RoutesDefinitionProvider from './providers/RoutesDefinitionProvider';
 import PartialSignatureHelpProvider from './providers/PartialSignatureHelpProvider';
 import CodeLensProvider from './providers/CodeLensProvider';
 import FormattingEditProvider from './providers/FormattingEditProvider';
-import { ViewCodeActionProvider, createPartialFromSelection } from './providers/ViewCodeActionProvider';
+import { ViewCodeActionProvider, createPartialFromSelection, wrapInConditional } from './providers/ViewCodeActionProvider';
 import DataAttributeCompletionProvider from './providers/DataAttributeCompletionProvider';
 import AssetsCompletionProvider from './providers/AssetsCompletionProvider';
 import ImagePreviewCodeLensProvider from './providers/ImagePreviewCodeLensProvider';
@@ -180,6 +180,11 @@ export class ExtensionActivator {
       vscode.commands.registerCommand(
         'hamlAll.createPartialFromSelection',
         createPartialFromSelection
+      ),
+
+      vscode.commands.registerCommand(
+        'hamlAll.wrapInConditional',
+        wrapInConditional
       ),
 
       vscode.commands.registerCommand(